requirements.txt
flask~=1.1.2
gunicorn
flask_sqlalchemy
flask_migrate
psycopg2
python-dotenv~=0.17.0
flask-login
SQLAlchemy~=1.4.8
Werkzeug~=1.0.1
alembic~=1.5.8
flask_cors
flask_script==2.0.6

requests~=2.24.0
python-dateutil~=2.8.1
